+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - BREAKING CHANGE changed character for `overdue` from `v` to `u`
 - Put single quotes around project names when printing them in terminal.
 - Add the ability to `project schedule` recurring tasks.
+- Add `skip-recurring` flag to `project schedule`.
 
 ## 2023-09-18 v0.5.2